// packages/mdd-engine/src/adapted-programs/programs/scaffold-file/types.ts
/**
 * Shared utility types for the scaffold-file script
 *
 * @noCanonicalDeclaration
 */

export type ScaffoldeeFileMetadata = {
  getImportStatement: (importConfiguration: ImportConfiguration) => string;
  camelCaseName: string;
  pascalCaseName: string;
  kebabCaseName: string;
  screamingSnakeCaseName: string;
};

export type ImportConfiguration = {
  filePath: string;
  identifierList: string[];
};